pyproject.toml
[project]
name = "railway-agent"
version = "4.0.0"
description = "工业化铁路出行智能Agent：分层架构 + 多轮对话 + 缓存 + 监控"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name" }
]
license = { text = "MIT" }
keywords = ["agent", "langchain", "langgraph", "railway", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "prometheus-client>=0.20.0",
    "cachetools>=5.0.0",
    "redis>=5.0.0",
    "tenacity>=8.0.0",
    "networkx>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "ipython>=8.0.0",
]

[project.scripts]
railway-agent = "cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
